requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
# Optional for Gemini integration:
# google-genai>=0.2.0
# Optional for CrewAI:
# crewai
